Move two IApply... interfaces to proper namespace
