+++ src/NUnitFramework/framework/Attributes/RequiresThreadAttribute.cs
@@ -24,6 +24,7 @@
 #if !SILVERLIGHT && !PORTABLE
 using System;
 using System.Threading;
+using NUnit.Framework.Interfaces;
 using NUnit.Framework.Internal;
 
 namespace NUnit.Framework
