+++ src/NUnitFramework/framework/Attributes/SetCultureAttribute.cs
@@ -23,6 +23,7 @@
 
 #if !NETCF
 using System;
+using NUnit.Framework.Interfaces;
 using NUnit.Framework.Internal;
 
 namespace NUnit.Framework
