+++ src/NUnitFramework/framework/Interfaces/IApplyToTest.cs
@@ -22,8 +22,9 @@
 // ***********************************************************************
 
 using System;
+using NUnit.Framework.Internal;
 
-namespace NUnit.Framework.Internal
+namespace NUnit.Framework.Interfaces
 {
     /// <summary>
     /// The IApplyToTest interface is implemented by self-applying
