+++ src/NUnitFramework/framework/Internal/Commands/ApplyChangesToContextCommand.cs
@@ -23,6 +23,8 @@
 
 using System;
 using System.Threading;
+using NUnit.Framework.Interfaces;
+
 namespace NUnit.Framework.Internal.Commands
 {
     /// <summary>
