+++ src/NUnitFramework/tests/Attributes/ApplyToContextTests.cs
@@ -22,6 +22,7 @@
 // ***********************************************************************
 
 using System.Globalization;
+using NUnit.Framework.Interfaces;
 using NUnit.Framework.Internal;
 
 namespace NUnit.Framework.Attributes
